reorganize darshan and provenance modules (move into `/tools`)
